refactor: station management 페이지 delete 로직 파일로 분리
